requirements.txt
bcrypt==4.0.1
fastapi==0.89.1
prisma==0.8.0
python-jose==3.3.0
uvicorn==0.20.0
strawberry-graphql==0.155.0
